Got table rendering to work. Removed itext project

+++ core/src/main/groovy/com/craigburke/document/core/Border.groovy
@@ -1,5 +1,5 @@
 package com.craigburke.document.core
 
 class Border implements AssignableColor {
-    BigDecimal size = 1
+    Integer size = 1
 }
