+++ core/src/main/groovy/com/craigburke/document/core/Cell.groovy
@@ -6,5 +6,5 @@ class Cell extends BaseNode {
 	Align align = Align.LEFT
 
 	Integer position = 0
-	BigDecimal width
+	Integer width
 }
