+++ core/src/main/groovy/com/craigburke/document/core/Font.groovy
@@ -2,7 +2,7 @@ package com.craigburke.document.core
 
 class Font implements AssignableColor, Cloneable {
 	String family = "Helvetica"
-	BigDecimal size = 12
+	Integer size = 12
 	Boolean bold = false
 	Boolean italic = false
